+++ src/solvers/global_optimisers/mod.rs
@@ -3,5 +3,3 @@ mod cross_entropy;
 
 pub use particle_swarm::ParticleSwarm;
 pub use cross_entropy::CrossEntropy;
-
-use super::VectorType;