+++ src/solvers/multivariable/gaussnewton_fd.rs
@@ -1,5 +1,4 @@
-use super::VectorType;
-use crate::{SolverError, SolverResult, DEFAULT_ITERMAX, DEFAULT_TOL};
+use crate::{SolverError, SolverResult, DEFAULT_ITERMAX, DEFAULT_TOL, VectorType};
 use nalgebra::{allocator::Allocator, ComplexField, DefaultAllocator, Dim, UniformNorm, U1};
 use num_traits::{Float, Signed};
 use std::marker::PhantomData;