+++ src/tests/mod.rs
@@ -1,4 +1,0 @@
-pub mod single;
-pub mod multi;
-pub mod ode;
-pub mod misc;